requirement.txt
Jinja2==2.8.1
MarkupSafe==2.0.1
PyYAML==6.0
